% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrc_methods.R
\name{get_clean_model}
\alias{get_clean_model}
\title{Get Clean LTRC Model Object}
\usage{
get_clean_model(mod)
}
\arguments{
\item{mod}{a model fit using the \code{ltrc()} function}
}
\value{
an object of class \code{"ltrc_mod"} containing clean information from the model fit, which can
then be passed to other methods
}
\description{
Get Clean LTRC Model Object
}
