% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrc.R
\name{ltrc}
\alias{ltrc}
\title{ltrc regression}
\usage{
ltrc(
  formula,
  trunc_time,
  data = NULL,
  n_start = 10,
  int_knots = 5,
  verbose = FALSE,
  n_folds = 5,
  knot_range = 1:10,
  knots = NULL,
  boundary_knots = NULL,
  sd_multiplier = 3,
  diag_only = TRUE,
  iter_max = 100,
  return_all = FALSE
)
}
\arguments{
\item{formula}{a \code{formula} object of the form \code{y ~ x}, where \code{y} must be a \code{Surv()} object
from the \code{survival} package}

\item{trunc_time}{a vector indicating the truncation time for each subject}

\item{data}{The data where the variables in the formula can be found}

\item{n_start}{the number of random starts for the fitting algorithm}

\item{int_knots}{the number of interior knots for the spline basis}

\item{verbose}{bool indicating if intermediate steps should be printed,
defaults to \code{FALSE}}

\item{n_folds}{the number of folds for cross-validation, if used}

\item{knot_range}{the number of knots to check for cross validation}

\item{knots}{if you want to fix the knots on the spline, pass the knots here}

\item{boundary_knots}{if you want to fix the knots on the spline, pass the boundary knots here}

\item{sd_multiplier}{controls the variance of the random noise for the parameter starting values}
}
\value{
a list with several elements
}
\description{
Left-truncated right-censored regression model fit using a seive maximum likelihood approach.
}
