% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrc_methods.R
\name{predict.ltrc_mod}
\alias{predict.ltrc_mod}
\title{Predict using an LTRC Model}
\usage{
\method{predict}{ltrc_mod}(object, newdata, ...)
}
\arguments{
\item{newdata}{a \code{matrix} or \code{data.frame} with the columns matching the order
of the predictors in the model}

\item{mod}{a model fit using the \code{ltrc()} function}
}
\value{
a vector of predictions
}
\description{
Predict using an LTRC Model
}
