% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrc_methods.R
\name{print.ltrc_mod}
\alias{print.ltrc_mod}
\title{Print an LTRC Model}
\usage{
\method{print}{ltrc_mod}(object, ...)
}
\arguments{
\item{object}{a model fit using the \code{ltrc()} function}
}
\value{
None
}
\description{
Prints out a brief summary of a fitted LTRC model.
}
