% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_iteration.R
\name{sim_iteration}
\alias{sim_iteration}
\title{Simulation Iteration}
\usage{
sim_iteration(
  i,
  FINAL_KNOTS,
  true_beta,
  n,
  error_dist,
  true_hazard,
  sd_multiplier = 3,
  n_start = 20,
  type_data = "ltrc",
  verbose = FALSE,
  diag_only = TRUE,
  return_all = FALSE
)
}
\arguments{
\item{i}{the number of iteration you are on}

\item{FINAL_KNOTS}{the number of knots used in the spline}

\item{true_beta}{the true beta vector}

\item{n}{the number of observations}

\item{error_dist}{a function for the error distribution}

\item{sd_multiplier}{the multiplier for the beta random starts}

\item{n_start}{the number of random starts in each iteration}

\item{type_data}{a character string, one of \code{c("ltrc", "lt", "rc", "none")}
signifying the type of data we are fitting the model to: left-truncated
right-censored, left-truncated only, right-censored only, or a normal linear model}
}
\value{
a list with the simulation results
}
\description{
Simulation Iteration
}
