% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_iteration_updated.R
\name{sim_iteration_updated}
\alias{sim_iteration_updated}
\title{Simulation Iteration Updated}
\usage{
sim_iteration_updated(
  i,
  FINAL_KNOTS,
  true_beta = c(0.5, 1),
  n,
  error_dist,
  sd_multiplier = 3,
  n_start = 20,
  type_data = "ltrc",
  verbose = FALSE,
  diag_only = TRUE,
  return_all = FALSE,
  C_UB = 44,
  sample_size_mult = 10
)
}
\arguments{
\item{i}{the number of iteration you are on}

\item{FINAL_KNOTS}{the number of knots used in the spline}

\item{true_beta}{the true beta vector, of size 3}

\item{n}{the number of observations}

\item{error_dist}{a function for the error distribution}

\item{sd_multiplier}{the multiplier for the beta random starts}

\item{n_start}{the number of random starts in each iteration}

\item{type_data}{a character string, one of \code{c("ltrc", "lt", "rc", "none")}
signifying the type of data we are fitting the model to: left-truncated
right-censored, left-truncated only, right-censored only, or a normal linear model}

\item{C_UB}{a numeric value representing the censoring distribution upper bound. For
50\% censoring = 11, for 30\% censoring = 17, for 15\% censoring = 30}

\item{sample_size_mult}{multiplier for the sample size (based on truncation prob)}
}
\value{
a list with the simulation results
}
\description{
Simulation Iteration Updated
}
