% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrc_methods.R
\name{summary.ltrc_mod}
\alias{summary.ltrc_mod}
\title{Summary of an LTRC Model}
\usage{
\method{summary}{ltrc_mod}(object, ...)
}
\arguments{
\item{mod}{a model fit using the \code{ltrc()} function}
}
\value{
None
}
\description{
Prints out a detailed summary of a fitted LTRC model.
}
