% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrc_methods.R
\name{vcov.ltrc_mod}
\alias{vcov.ltrc_mod}
\title{Variance-Covariance Matrix for LTRC Model Parameters}
\usage{
\method{vcov}{ltrc_mod}(
  object,
  type = c("efficient_score", "inverse_information"),
  spline_params = FALSE,
  ...
)
}
\arguments{
\item{object}{a model fit using the \code{ltrc()} function}

\item{type}{character string specifying the method used to compute the
variance, one of \code{c("efficient_score", "inverse_information")}}

\item{spline_params}{Logical indicating if the spline parameter variances
should be included in the returned matrix. Only for \verb{type = "inverse_information}.
Defaults to \code{FALSE}.}

\item{...}{unused additional arguments}
}
\value{
a variance-covariance matrix
}
\description{
Variance-Covariance Matrix for LTRC Model Parameters
}
